feat: extract common files to shared folder
