+++ examples/demo/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "../../tsconfig.base.json",
   "files": [],
-  "include": [],
+  "include": ["../../shared/*.ts"],
   "references": [
     {
       "path": "./tsconfig.app.json"
