+++ examples/shared/environment.ts
