+++ examples/shared/index.ts
@@ -1,0 +1,9 @@
+// start:ng42.barrel
+export * from './todos.animation';
+export * from './todos.config';
+export * from './todos.model';
+export * from './todos.replication';
+export * from './todos.migration';
+export * from './environment';
+// end:ng42.barrel
+
