+++ examples/shared/todos.animation.ts
@@ -1,0 +1,17 @@
+import { animate, query, stagger, style, transition, trigger } from '@angular/animations';
+
+export const todosListAnimation = trigger('listAnimation', [
+  transition('* <=> *', [
+    query(
+      ':enter',
+      [
+        style({ opacity: 0 }),
+        stagger('50ms', animate('60ms ease-in', style({ opacity: 1 }))),
+      ],
+      { optional: true }
+    ),
+    query(':leave', stagger('10ms', animate('50ms ease-out', style({ opacity: 0 }))), {
+      optional: true,
+    }),
+  ]),
+]);
