+++ package.json
@@ -120,7 +120,6 @@
     "@swc-node/register": "~1.6.7",
     "@swc/core": "~1.3.85",
     "@twittwer/compodoc": "1.10.0",
-    "@types/dotenv": "8.2.0",
     "@types/fs-extra": "^11.0.2",
     "@types/jest": "^29.4.0",
     "@types/node": "18.16.9",
